Remove failed tasks from inflightRequests array
