+++ Sources/MapboxNavigation/ImageDownloader/ImageDownloaderProtocol.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol ImageDownloaderProtocol {
+    func download(with url: URL, completion: @escaping (Result<CachedURLResponse, Error>) -> Void)
+}
